Fixed (?) play button issues
Changed function back to "play" - left Benas's playMusic in case we need
it in the future - and made some changes so that sounds play when play
is pressed but not when the button is pressed.
For some reason the sounds each grid square represented were flipped
when they were touched as opposed to played with the play button, so I
made those consistent.