threaded code working; start on animation
Added threading and TimerTask so that the view stuff goes on the UI
thread soundpool executes on a separate thread.

There's a new class, ImmutBooleanGridModel, which may or may not be
necessary. I'm committing it so we all have it to hand, but I think -
/think/ - that we're okay without it at the moment.

Animation is in pre-alpha. will not de-select and will never select the
first column. Should only 'work' for one grid, behavior with more than
one may get wacky.