plays sound on touch, updates while playing
Two-line fix to make it update the grid during playback. Also
implemented the play-sound-on-touch option, which is a bit weird if you
do the edit while it's playing back. If you lot are interested, we could
add in a logic loop so it only plays the sound if it's selected while
the song isn't already looping.